[1/n] ensure uniqueness of fake artifact hashes
In Omicron we're adding a constraint that (deployment unit kind, hash) is
globally unique. We can't ensure *global* uniqueness in tufaceous at the moment,
but we can ensure uniqueness within a repository.

This change makes it so that deployment units are unique, but does not verify
this. The next commit in the series (#16) will perform this verification (that's a
bit big so I wanted to separate this out).

Reviewers: iliana

Reviewed By: iliana

Pull Request: https://github.com/oxidecomputer/tufaceous/pull/15