[2/n] add tests for duplicate control plane zones and artifacts
We already have code which ensures that duplicate zones and artifacts are detected. Add tests to exercise these code paths (we're going to hoist them to a higher level in a subsequent commit).

While writing this code, I realized that we don't exit with code 0 when encountering duplicate artifacts. That's not good. I'll fix it in a subsequent commit.

Reviewers: iliana

Reviewed By: iliana

Pull Request: https://github.com/oxidecomputer/tufaceous/pull/18