[4/n] verify uniqueness of kind/hash for deployment units
This ended up being a bit of bigger PR -- it verifies that when constructing a new repo, the deployment units are all unique (within the scope of the repo).

We previously had some code which verified this for control plane zones -- hoist that out into a common location and use it to verify the invariant at the repository scope too. (We could also extend the uniqueness check to work for name/version/kind as well, but I've left that for future work.)

The `artifacts.json` format isn't changed. I think ultimately if we do want to have a serialized deployment unit manifest, it should live inside the artifact (so it's self-describing) rather than outside it.

One challenge here is that while working with existing repos, or with tarballs from an external build process, a proper implementation would extract those tarballs out to obtain the deployment units. That code lives in Omicron, however. In the interest of urgency, we don't do this verification in those cases, and we just have an `Unknown` field for now. (What we want to do here in the end will be determined by the outcome of RFD 554.)

Reviewers: iliana, jgallagher

Reviewed By: jgallagher

Pull Request: https://github.com/oxidecomputer/tufaceous/pull/16