preserve requested formatting in some Display impls (#22)
I'm writing some tooling with this and noticed that when I formatted
`ArtifactVersion` and `ArtifactKind` into fixed-width fields, the width
wasn't being honored. I guess this is the way you're supposed to
delegate `Display`? Here's a playground that shows the problem outside
of tufaceous:

https://play.rust-lang.org/?version=stable&mode=debug&edition=2024&gist=80fe4b1d5e37fb96f2abd500722bc36a

There are other impls in tufaceous that I didn't fix because they
weren't in my way and it was not obvious how best to do it.