repository load methods with trusted roots (#30)
Part of https://github.com/oxidecomputer/omicron/issues/3578.

The current plan is to put our trust store in CockroachDB, which allows
us to have multiple trust roots uploaded, so we loop through the roots
in the trust store and try to load the repository metadata with each
one. If we fail to verify any of the metadata, we try the next root in
the provided trust store.