remove code to add artifacts incrementally (#38)
All of our production users construct and then assemble a manifest, and
that approach has proven to be easier to work with than incremental
additions. Significant parts of the code can be made simpler this way,
since they no longer need to construct a view of what's in the TUF repo
by parsing the files contained in them.

The tufaceous binary now consists of just the assemble and extract
commands. (We still use the binary in some tests, and it is convenient
to have in general.)

I patched this change into Omicron and verified that it built and passed
update-related tests successfully.