Make fake BORD and SIGN values a little more realistic (#40)
When trying Reconfigurator-based updates on a racklette today with a
real TUF repo, the planner skipped all the RoT bootloader and RoT
updates because none of the artifact `name`s matched the artifact
`board`s. Artifact name matching board is a property that's true for the
SP in both fake data and production, but is only true for the bootloader
and RoT in the fake data generated by tufaceous.

This PR changes the fake data to be more consistent with real data:

* TUF artifact `name` does not necessarily match hubris caboose `name`
* All RoTs and RoT bootloaders report the same `BORD`
* Each target type (sled, switch, psc) has distinct `SIGN` values
despite the `BORD` values matching